advanced_workflow/pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
description = "AI-powered news aggregation system that collects content from multiple sources, generates daily LLM-powered summaries, and delivers personalized digests via email"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "markdown>=3.7.0",
    "markdownify>=0.11.6",
    "openai>=2.7.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "youtube-transcript-api>=1.2.3",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=7.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
